move util function related to downgrades into separate downgrades file
